Dropbox implementation
As you will see their are a few problems with the DropBoxChooser. First issue
is that it will always trigger an error in dropbox.ks:208 because the popup
isn't detected. Therefore we have a second issue: the popup can't be closed
by the code. And last: the cancel-button is useless.